+++ src/Controller/SecurityController.php
@@ -6,6 +6,8 @@ use App\Entity\User\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use JsonException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -32,15 +34,11 @@ class SecurityController extends AbstractController
     }
 
     /**
-     * @Route("/login", name="app_login")
+     * Login page controller
      */
-    #[Route('/login', name: 'app_login')]
+    #[Route('/login', name: 'login')]
     public function login(AuthenticationUtils $authenticationUtils): Response
     {
-        // if ($this->getUser()) {
-        //     return $this->redirectToRoute('target_path');
-        // }
-
         // get the login error if there is one
         $error = $authenticationUtils->getLastAuthenticationError();
         // last username entered by the user
@@ -91,10 +89,31 @@ class SecurityController extends AbstractController
             'sign_in_label' => 'Log in',
 
             // the 'name' HTML attribute of the <input> used for the username field (default: '_username')
-            'username_parameter' => 'username',
+            'username_parameter' => '_iam_username',
 
             // the 'name' HTML attribute of the <input> used for the password field (default: '_password')
-            'password_parameter' => 'password',
+            'password_parameter' => '_iam_password',
+
+            // whether to enable or not the "forgot password?" link (default: false)
+            'forgot_password_enabled' => false,
+
+            // the path (i.e. a relative or absolute URL) to visit when clicking the "forgot password?" link (default: '#')
+            // 'forgot_password_path' => $this->generateUrl('...', ['...' => '...']),
+
+            // the label displayed for the "forgot password?" link (the |trans filter is applied to it)
+            'forgot_password_label' => 'Forgot your password?',
+
+            // whether to enable or not the "remember me" checkbox (default: false)
+            'remember_me_enabled' => true,
+
+            // remember me name form field (default: '_remember_me')
+            'remember_me_parameter' => '_iam_remember_me',
+
+            // whether to check by default the "remember me" checkbox (default: false)
+            'remember_me_checked' => false,
+
+            // the label displayed for the remember me checkbox (the |trans filter is applied to it)
+            'remember_me_label' => 'Remember me',
         ]);
     }
 
@@ -114,6 +133,7 @@ class SecurityController extends AbstractController
      * @throws JsonException
      */
     #[Route('/api/change_user_password', name: 'app_change_password_old', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function change_password_old(Request $request,
                                         UserPasswordHasherInterface $passwordHasher): JsonResponse
     {
@@ -127,6 +147,7 @@ class SecurityController extends AbstractController
      * @throws JsonException
      */
     #[Route('/api/users/change_password', name: 'app_change_password', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function change_password(Request $request,
                                     UserPasswordHasherInterface $passwordHasher): JsonResponse
     {
@@ -152,7 +173,7 @@ class SecurityController extends AbstractController
 
         // Do a cross check so only the user can change their password
         $currentUser = $this->getUser();
-        if (null === $currentUser || $currentUser->getUsername() !== $user->getUsername()) {
+        if (null === $currentUser || $currentUser->getUserIdentifier() !== $user->getUserIdentifier()) {
             return $this->json([
                 'code' => 401,
                 'error' => 'Invalid token access.'
@@ -185,6 +206,7 @@ class SecurityController extends AbstractController
      * @throws JsonException
      */
     #[Route('/api/change_password_by_sikka', name: 'app_change_password_by_sikka_old', methods: ['POST'])]
+    #[Security("is_granted('ROLE_SUPER_ADMIN') or is_granted('ROLE_HRIS') or is_granted('ROLE_UPK_PUSAT') or is_granted('ROLE_UPK_WILAYAH') or is_granted('ROLE_UPK_LOKAL')")]
     public function change_password_by_sikka_old(Request $request,
                                                  UserPasswordHasherInterface $passwordHasher): JsonResponse
     {
@@ -198,6 +220,7 @@ class SecurityController extends AbstractController
      * @throws JsonException
      */
     #[Route('/api/users/change_password_by_sikka', name: 'app_change_password_by_sikka', methods: ['POST'])]
+    #[Security("is_granted('ROLE_SUPER_ADMIN') or is_granted('ROLE_HRIS') or is_granted('ROLE_UPK_PUSAT') or is_granted('ROLE_UPK_WILAYAH') or is_granted('ROLE_UPK_LOKAL')")]
     public function change_password_by_sikka(Request $request,
                                              UserPasswordHasherInterface $passwordHasher): JsonResponse
     {
@@ -243,6 +266,7 @@ class SecurityController extends AbstractController
      * @return JsonResponse
      */
     #[Route('/api/whoami', name: 'app_whoami_old', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function whoamiOld(): JsonResponse
     {
         return $this->json($this->getUser());
@@ -252,6 +276,7 @@ class SecurityController extends AbstractController
      * @return JsonResponse
      */
     #[Route('/api/token/whoami', name: 'app_whoami', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function whoami(): JsonResponse
     {
         return $this->json($this->getUser());
@@ -263,6 +288,7 @@ class SecurityController extends AbstractController
      * @throws JsonException
      */
     #[Route('/api/users/check_identifier', name: 'app_check_user_identifier', methods: ['POST'])]
+    #[IsGranted('PUBLIC_ACCESS')]
     public function checkValidUserIdentifier(Request $request): JsonResponse
     {
         $content = json_decode($request->getContent(), true, 512, JSON_THROW_ON_ERROR);