+++ src/OpenApi/AuthenticationDecorator.php
@@ -12,7 +12,7 @@ use ArrayObject;
 final class AuthenticationDecorator implements OpenApiFactoryInterface
 {
     public function __construct(
-        private OpenApiFactoryInterface $decorated
+        private readonly OpenApiFactoryInterface $decorated
     ) {}
 
     public function __invoke(array $context = []): OpenApi